DL-17828 | Change Identifier type from Int to Long

+++ app/uk/gov/hmrc/pla/stub/model/hip/HIPProtectionsModel.scala
@@ -53,7 +53,7 @@ object ProtectionRecordsList {
 }
 
 case class ProtectionRecord(
-    identifier: Int,
+    identifier: Long,
     sequenceNumber: Int,
     `type`: ProtectionType,
     certificateDate: String,
@@ -79,7 +79,7 @@ object ProtectionRecord {
 
   def apply(protection: Protection): ProtectionRecord =
     ProtectionRecord(
-      protection.id.toInt,
+      protection.id,
       protection.version,
       toRecordType(protection),
       protection.certificateDate.getOrElse("2000-01-01"), // TODO: defaults for date and time?
